Do not source /* if BASH_COMPLETION_COMPAT_DIR is unset
If BASH_COMPLETION_COMPAT_DIR was unset it tried to source /*
wich lead to repl not showing up if there were files in / that
lead no termination of sourcing. This could potentially do harm
if there is a harmful script in /.

$BASH_COMPLETION_DEFAULT_DIR/* will now only be sourced if
BASH_COMPLETION_DEFAULT_DIR is set and is not empty.